pyproject.toml
[project]
name = "storefront"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "django-templated-mail>=1.1.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-nested-routers>=0.95.0",
    "flower>=2.0.1",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "redis>=7.2.1",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "django-debug-toolbar>=6.2.0",
    "django-silk>=5.4.3",
    "locust>=2.43.3",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
